requirements.txt
streamlit
numpy
opencv-python-headless
pillow
torch
torchvision
pymysql
matplotlib
